+++ internal/domain/entities/product.go
@@ -5,6 +5,6 @@ import "github.com/google/uuid"
 type Product struct {
 	ID      uuid.UUID
 	Name    string
-	Price   float64
+	Price   int
 	BrandID uuid.UUID
 }
