+++ internal/domain/entities/voucher.go
@@ -12,6 +12,6 @@ type Voucher struct {
 	CostInPoint int       `json:"cost_in_point"`
 	Expiration  time.Time `json:"expiration"`
 	Type        string    `json:"type"`
-	Value       float64   `json:"value"`
+	Value       int       `json:"value"`
 	BrandID     uuid.UUID `json:"brand_id"`
 }
